add README + clean up + refactor fns out of main
