+++ requirements.txt
@@ -1,0 +1,42 @@
+# This file was autogenerated by uv via the following command:
+#    uv pip compile pyproject.toml
+annotated-types==0.7.0
+    # via pydantic
+certifi==2025.1.31
+    # via requests
+charset-normalizer==3.4.1
+    # via requests
+dotenv==0.9.9
+    # via mbta-challenge (pyproject.toml)
+exceptiongroup==1.2.2
+    # via pytest
+idna==3.10
+    # via requests
+iniconfig==2.0.0
+    # via pytest
+packaging==24.2
+    # via pytest
+pluggy==1.5.0
+    # via pytest
+pydantic==2.10.6
+    # via mbta-challenge (pyproject.toml)
+pydantic-core==2.27.2
+    # via pydantic
+pytest==8.3.5
+    # via mbta-challenge (pyproject.toml)
+python-dotenv==1.0.1
+    # via dotenv
+requests==2.32.3
+    # via mbta-challenge (pyproject.toml)
+tomli==2.2.1
+    # via pytest
+types-requests==2.32.0.20250306
+    # via mbta-challenge (pyproject.toml)
+typing-extensions==4.12.2
+    # via
+    #   pydantic
+    #   pydantic-core
+urllib3==2.3.0
+    # via
+    #   requests
+    #   types-requests
